graphab/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2025.1.31
cffi==1.17.1
GDAL==3.10.3
markdown-it-py==3.0.0
mdurl==0.1.2
minio==7.2.15
numpy==2.2.4
pandas==2.2.3
pycparser==2.22
pycryptodome==3.22.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
rich==14.0.0
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
matplotlib==3.10.1
PyYAML==6.0.2